==> autoencoder_4_microbiota/00_cMD_Counts_DataOverview.ipynb <==
import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt


df_counts = pd.read_csv("../data/taxa_data_counts.csv", sep=',', header=0, index_col=0)
df_counts.columns = df_counts.columns.str.replace('[', '').str.replace(']', '')
df_counts = df_counts.fillna(0)
df_counts.insert(0, 'total_counts', df_counts.sum(axis=1))
df_counts

import matplotlib.pyplot as plt
import seaborn as sns 
plt.figure(figsize=(12, 4))
sns.histplot(data=df_counts[df_counts.total_counts < 5e6], x='total_counts', bins=100)
plt.title('Distribution of Total Counts in Samples (zoom in samples total_counts < 5e6)')

import matplotlib.pyplot as plt
import seaborn as sns

# Create bins of 10 million
bins = range(0, 200000001, 10000000)
df_counts['binned'] = pd.cut(df_counts['total_counts'], bins=bins)

# Count the number of occurrences in each bin
bin_counts = df_counts['binned'].value_counts().sort_index()

# Calculate percentages
bin_percentages = (bin_counts / bin_counts.sum()) * 100

# Create a DataFrame for visualization
bin_df = pd.DataFrame({'count': bin_counts, 'percentage': bin_percentages})

# Plot the counts and percentages
plt.figure(figsize=(12, 6))
ax = sns.barplot(x=bin_df.index.astype(str), y=bin_df['count'], palette='viridis')
ax.set_xticklabels(ax.get_xticklabels(), rotation=90)
plt.title('Counts of Total Counts by Bins of 10 Million (First 20 Bins)')
plt.xlabel('Total Counts Bins')
plt.ylabel('Counts')

# Annotate the bars with counts and percentages
for p in ax.patches:
    height = p.get_height()
    ax.annotate(f'{int(height)}\n({height / bin_counts.sum() * 100:.2f}%)', 
                (p.get_x() + p.get_width() / 2., height), 
                ha='center', va='center', 
                xytext=(0, 10), 
                textcoords='offset points')

plt.show()
